+++ front/src/styles/components/CourseCard.css
