+++ front/src/styles/components/CourseList.css
