+++ front/src/styles/components/NavigationBar.css
