+++ front/src/styles/pages/CoursesPage.css
@@ -4,7 +4,7 @@ body {
 }
 
 .background-img {
-    background-image: url("./images/chalkboard.jpg");
+    background-image: url("../../images/chalkboard.jpg");
     width: 100%;
     height: 100%;
     position: fixed;
